--- src/main/java/com/opencode/alumxbackend/users/model/Chat.java
@@ -1,34 +1,0 @@
-package com.opencode.alumxbackend.users.model;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "chats")
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Builder
-public class Chat {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(columnDefinition = "BIGINT")
-    private Long id;
-    @Column(nullable = false)
-    private String user1Username;
-    @Column(nullable = false)
-    private String user2Username;
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @PrePersist
-    protected void onCreate() {
-        createdAt = LocalDateTime.now();
-    }
-}
