--- src/main/java/com/opencode/alumxbackend/users/model/Message.java
@@ -1,38 +1,0 @@
-package com.opencode.alumxbackend.users.model;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@Table(name = "messages")
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Builder
-public class Message {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(columnDefinition = "BIGINT")
-    private Long id;
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_id",nullable = false)
-    private Chat chat;
-    @Column(nullable = false)
-    private String senderUsername;
-    @Column(nullable = false)
-    private String receiverUsername;
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String content;
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @PrePersist
-    protected void onCreate(){
-        createdAt = LocalDateTime.now();
-    }
-}
